requirements.txt
asgiref==3.5.2
autopep8==2.0.0
Django==4.1.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
psycopg2==2.9.5
pycodestyle==2.9.1
PyJWT==1.7.1
python-dotenv==0.21.0
pytz==2022.4
sqlparse==0.4.3
tomli==2.0.1
